Implement #[property(rpc = "mode")]
