+++ gdnative-derive/src/syntax.rs
@@ -1,0 +1,1 @@
+pub mod rpc_mode;
